feat: Add passthrough functions
