+++ cache.go
@@ -2,9 +2,7 @@ package sturdyc
 
 import (
 	"context"
-	"errors"
 	"hash/fnv"
-	"maps"
 	"sync"
 	"time"
 )
@@ -54,6 +52,9 @@ type Client struct {
 	bufferPermutationIDs  map[string][]string
 	bufferPermutationChan map[string]chan<- []string
 
+	passthroughPercentage int
+	passthroughBuffering  bool
+
 	useRelativeTimeKeyFormat bool
 	keyTruncation            time.Duration
 }
@@ -71,9 +72,10 @@ func New(capacity, numShards int, ttl time.Duration, evictionPercentage int, opt
 	// Create a new client, and apply the options.
 	//nolint: exhaustruct // The options are going to set the remaining fields.
 	client := &Client{
-		ttl:              ttl,
-		clock:            NewClock(),
-		evictionInterval: ttl / time.Duration(numShards),
+		ttl:                   ttl,
+		clock:                 NewClock(),
+		evictionInterval:      ttl / time.Duration(numShards),
+		passthroughPercentage: 100,
 	}
 
 	for _, opt := range opts {
@@ -113,6 +115,12 @@ func (c *Client) Size() int {
 		sum += shard.size()
 	}
 	return sum
+}
+
+// Delete removes a single entry from the cache.
+func (c *Client) Delete(key string) {
+	shard := c.getShard(key)
+	shard.delete(key)
 }
 
 // startEvictions is going to be running in a separate goroutine that we're going to prevent from ever exiting.
@@ -158,176 +166,4 @@ func (c *Client) reportCacheHits(cacheHit bool) {
 func (c *Client) set(key string, value any, isMissingRecord bool) bool {
 	shard := c.getShard(key)
 	return shard.set(key, value, isMissingRecord)
-}
-
-// get retrieves a value from the cache and performs a type assertion to the desired type.
-func get[T any](c *Client, key string) (value T, exists, ignore, refresh bool) {
-	shard := c.getShard(key)
-	entry, exists, ignore, refresh := shard.get(key)
-	c.reportCacheHits(exists)
-
-	if !exists {
-		return value, false, false, false
-	}
-
-	val, ok := entry.(T)
-	if !ok {
-		return value, false, false, false
-	}
-
-	return val, exists, ignore, refresh
-}
-
-// Get retrieves a value from the cache and performs a type assertion to the desired type.
-func Get[T any](c *Client, key string) (T, bool) {
-	value, ok, _, _ := get[T](c, key)
-	return value, ok
-}
-
-// GetFetch attempts to retrieve the specified key from the cache. If the value
-// is absent, it invokes the "fetchFn" to obtain it, and subsequently stores
-// it. Additionally, when stampede protection is active, GetFetch assesses
-// whether the record requires refreshing and, if so, schedules this task to be
-// executed in the background.
-
-// GetFetch attempts to retrieve the specified key from the cache. If the value
-// is absent, it invokes the "fetchFn" function to obtain it and then stores
-// the result. Additionally, when stampede protection is enabled, GetFetch
-// determines if the record needs refreshing and, if necessary, schedules this
-// task for background execution.
-func GetFetch[T any](ctx context.Context, c *Client, key string, fetchFn FetchFn[T]) (T, error) {
-	// Begin by checking if we have the item in our cache.
-	value, ok, shouldIgnore, shouldRefresh := get[T](c, key)
-
-	// We have the item cached, and we'll check if it should be refreshed in the background.
-	if shouldRefresh {
-		safeGo(func() {
-			refresh(c, key, fetchFn)
-		})
-	}
-
-	if shouldIgnore {
-		return value, ErrMissingRecord
-	}
-
-	// If we don't have this item in our cache, we'll fetch it
-	if !ok {
-		response, err := fetchFn(ctx)
-		if err != nil {
-			// In case of an error, we'll only cache the response if the fetchFn returned an ErrMissingRecord.
-			if c.storeMisses && errors.Is(err, ErrStoreMissingRecord) {
-				c.set(key, response, true)
-			}
-			return response, err
-		}
-
-		// Cache the response
-		c.set(key, response, false)
-		return response, err
-	}
-
-	return value, nil
-}
-
-// GetFetchBatch attempts to retrieve the specified ids from the cache. If any
-// of the values are absent, it invokes the fetchFn function to obtain them and
-// then stores the result. Additionally, when stampede protection is enabled,
-// GetFetch determines if any of the records needs refreshing and, if
-// necessary, schedules this to be performed in the background.
-func GetFetchBatch[T any](
-	ctx context.Context,
-	c *Client,
-	ids []string,
-	keyFn KeyFn,
-	fetchFn BatchFetchFn[T],
-) (map[string]T, error) {
-	cachedRecords := make(map[string]T)
-	cacheMisses := make([]string, 0)
-	idsToRefresh := make([]string, 0)
-
-	// Group the IDs into cached records, cache misses, and ids which are due for a refresh.
-	for _, id := range ids {
-		key := keyFn(id)
-		value, exists, shouldIgnore, shouldRefresh := get[T](c, key)
-
-		// Check if the record should be refreshed in the background.
-		if shouldRefresh {
-			idsToRefresh = append(idsToRefresh, id)
-		}
-
-		if shouldIgnore {
-			continue
-		}
-
-		if !exists {
-			cacheMisses = append(cacheMisses, id)
-			continue
-		}
-
-		cachedRecords[id] = value
-	}
-
-	// If any records need to be refreshed, we'll do so in the background.
-	if len(idsToRefresh) > 0 {
-		if c.bufferRefreshes {
-			safeGo(func() {
-				bufferBatchRefresh(c, idsToRefresh, keyFn, fetchFn)
-			})
-		} else {
-			safeGo(func() {
-				refreshBatch(c, idsToRefresh, keyFn, fetchFn)
-			})
-		}
-	}
-
-	// If we were able to retrieve all records from the cache, we can return them straight away.
-	if len(cacheMisses) == 0 {
-		return cachedRecords, nil
-	}
-
-	// Fetch the missing records.
-	response, err := fetchFn(ctx, cacheMisses)
-	if err != nil {
-		// We had some records in the cache, but the remaining records couldn't be
-		// retrieved. Therefore, we'll return the cached records along with a
-		// ErrOnlyCachedRecords error, and let the caller decide what to do.
-		if len(cachedRecords) > 0 {
-			return cachedRecords, ErrOnlyCachedRecords
-		}
-		return cachedRecords, err
-	}
-
-	// Check if we should store any of these IDs as a missing record.
-	if c.storeMisses && len(response) < len(cacheMisses) {
-		for _, id := range cacheMisses {
-			if v, ok := response[id]; !ok {
-				c.set(keyFn(id), v, true)
-			}
-		}
-	}
-
-	// Cache the fetched records, and merge them with the ones we already had.
-	for id, record := range response {
-		c.set(keyFn(id), record, false)
-	}
-	maps.Copy(cachedRecords, response)
-
-	return cachedRecords, nil
-}
-
-// Set writes a single value to the cache. Returns true if it triggered an eviction.
-func Set(c *Client, key string, value any) bool {
-	return c.set(key, value, false)
-}
-
-// SetMany writes multiple values to the cache. Returns true if it triggered an eviction.
-func SetMany[T any](c *Client, records map[string]T, cacheKeyFn KeyFn) bool {
-	var triggeredEviction bool
-	for id, value := range records {
-		evicted := c.set(cacheKeyFn(id), value, false)
-		if evicted {
-			triggeredEviction = true
-		}
-	}
-	return triggeredEviction
 }
